MVP, functionality is changing the colours of spot
A small MVP has been created, it gives the assessment page SPOT some more character, and allows for customization. Currently that customization is not saved, so in most cases the screen will be what is hard coded, unless you go into the extension and write in your color. Notes for future development: add a way to save the theme that the user creates